pyproject.toml
[tool.poetry]
name = "messenger"
version = "0.1.0"
description = ""
authors = ["vadjango <z9661744@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
channels = "^4.0.0"
channels-redis = "^4.1.0"
djangorestframework = "^3.14.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.1.0"}
pillow = "^10.0.0"
djangorestframework-simplejwt = "^5.2.2"
daphne = "^4.0.0"
django-extensions = "^3.2.3"
ipython = "^8.14.0"
twilio = "^8.7.0"
psycopg2-binary = "^2.9.7"
pyotp = "^2.9.0"
django-cors-headers = "^4.2.0"
django = "^4.2.5"
mypy = "^1.5.1"
djangorestframework-stubs = "^3.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
